+++ Python/Data-Structures/Stack/stack.py
